+++ NEWS.md
@@ -1,7 +1,18 @@
+# cubeview 0.2.4 (2025-04-12)
+
+#### 💬 documentation etc
+
+* add proper package anchors as requested by CRAN.
+
+## cubeview 0.2.2
+
+* legend now uses less space.
+* include argument options (cubeOptions()) to specify initial location of slices.
+
 ## cubeview 0.2.0
 
-* internal processing now much faster
+* internal processing now much faster.
 
 ## cubeview 0.1.0
 
-* Initial release
+* Initial release.